Add Flyweight pattern
The concept of this is World has grid of fixed size 10x10.
Player has limited number of steps to move around within the grid
to try to reach the destination which there are only 2 slots in the grid.

Each slot (or Tile) has cost of movement. If player can reach the destination
then win, otherwise lose.